Move getTime()-related stuff to gettime.h
