+++ src/threading/thread.cpp
@@ -30,17 +30,16 @@ DEALINGS IN THE SOFTWARE.
 
 #define UNUSED(expr) do { (void)(expr); } while (0)
 
+#include <cassert>
 #if USE_CPP11_THREADS
 	#include <chrono>
 	#include <system_error>
-	#include <cassert>
 #elif USE_WIN_THREADS
 	#ifndef _WIN32_WCE
 		#include <process.h>
 	#endif
 #elif USE_POSIX_THREADS
 	#include <time.h>
-	#include <assert.h>
 	#include <stdlib.h>
 	#include <unistd.h>
 	#include <sys/time.h>
